// frontend/src/utils/types.ts
export type submittedDataTypeIndividual = {
  firstName: string;
  lastName: string;
  email: string;
  country: string;
  location: string;
  channel: string;
};

export type submittedDataTypeBusiness = {
  businessName: string;
  email: string;
  organizationCountry: string;
  organizationSector: string;
  channel: string;
};